refactor: refactor game.js Refactor game.js so similar code can be used in search.js
