+++ src/components/Filter/FilterButtons/FilterButtons.js
@@ -1,0 +1,26 @@
+import React, { Fragment } from "react";
+
+// import "./LevelFilter.css";
+const FilterButtons = ({
+  setFilterOpen,
+  filterSet,
+  setFilterValue,
+  setOptions,
+}) => {
+  return (
+    <Fragment>
+      <div>
+        <button className="button-reset" onClick={() => setFilterOpen(true)}>
+          Filter
+        </button>
+        <button
+          className="button-filter-clear"
+          onClick={() => filterSet({}, setFilterValue, setOptions)}
+        >
+          Clear Filter
+        </button>
+      </div>
+    </Fragment>
+  );
+};
+export default FilterButtons;
